get rests from server
