+++ client/package.json
@@ -23,6 +23,7 @@
     "lodash": "^4.17.21",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-query": "^3.39.0",
     "react-router": "^6.2.1",
     "react-router-dom": "^6.2.2",
     "style-loader": "^3.3.1",
