+++ client/src/components/App.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
+
 import AppBar from './AppBar/AppBar';
 import RequireAuth from "../auth/RequireAuth";
