+++ client/src/types/Resturants.ts
@@ -1,5 +1,20 @@
 interface Restaurant {
-id: number;
-name: string;
-picture: string;
+    id: string;
+    name: string;
+    description: string;
+    imgUrl: string;
+    tags: string[];
+    openingTimes: {
+        [day: string]: string[];
+    //   "1": string,
+    //   "2": [{ type: String }],
+    //   "3": [{ type: String }],
+    //   "4": [{ type: String }],
+    //   "5": [{ type: String }],
+    },
+    location:  string;
+    contactIndo: {
+      phoneNumber:  string;
+      email: string;
+    }  
 }
